refactor: split dialogue and combat helpers
