+++ scripts/combat_ui_helpers.js
@@ -1,0 +1,3 @@
+export function updateHpBar(bar, current, max) {
+  bar.style.width = `${(current / max) * 100}%`;
+}
